refactor: 피드백 반영, 오류 수정 * return type (int -> double)

+++ src/main/java/lotto/domain/Result.java
@@ -19,7 +19,7 @@ public class Result {
     }
 
     public double getTotalWinningPrice() {
-        return result.stream().mapToInt(rank -> rank.getWinningMoney()).sum();
+        return result.stream().mapToDouble(rank -> rank.getWinningMoney()).sum();
     }
 
     public String getWinningRatio(long purchasedPrice) {
